+++ manage.py
@@ -1,5 +1,6 @@
-#!/usr/bin/env python
-"""Django's command-line utility for administrative tasks."""
+#!/bin/bash
+"exec" "`dirname $0`/../venv/bin/python" "$0" "$@"  # NOQA
+
 import os
 import sys