+++ terra_accounts/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
-from django.conf import settings
+from terra_settings import settings
 from rest_framework_jwt.settings import api_settings
 
 from .permissions_mixins import PermissionRegistrationMixin
@@ -15,7 +15,8 @@ class AccountsConfig(PermissionRegistrationMixin, AppConfig):
     )
 
     def ready(self):
-        # terra_accounts need to add jwt_delta key in terra_utils settings endpoint
+        # terra_accounts need to add jwt_delta key in terra-settings settings endpoint
         terra_settings = getattr(settings, 'TERRA_APPLIANCE_SETTINGS', {})
         terra_settings.setdefault('jwt_delta', api_settings.JWT_EXPIRATION_DELTA)
+        settings.TERRA_SETTINGS = terra_settings
         super().ready()